+++ application/controllers/Parts.php
@@ -7,8 +7,10 @@ class Parts extends Application
 	function __construct()
 	{
 		parent::__construct();
-		$this->load->model('PartsModel');
+		$this->load->model('partsmodel');
+        $this->load->model('apikeymodel');        
 	}
+
 	public function index()
 	{
             $role = $this->session->userdata('userrole');
@@ -19,8 +21,7 @@ class Parts extends Application
                 // Sets the view body of the page
                 $this->data['pagebody'] ='Parts';        
                 // Retrive all parts from model
-                $source = $this->PartsModel->all();     
-
+                $source = $this->partsmodel->all();    
                 // Loops through all the parts and addds them to the array
                 foreach($source as $record){
                     $record['part_pic'] = $record['part_code'] . ".jpeg";
@@ -46,5 +47,83 @@ class Parts extends Application
             else{
                 redirect('Welcome');
             }
+        }	
+
+    public function build(){
+        // Get API key from database
+        $key = $this->apikeymodel->getKey()[0]['apikey'];
+
+        if($key != null){
+            $arrContextOptions=array(
+            "ssl"=>array(
+                "verify_peer"=>false,
+                "verify_peer_name"=>false,
+                ),
+            ); 
+
+            // Get json response from REST API
+            $response = file_get_contents('https://umbrella.jlparry.com/work/mybuilds?key=' . $key, false, stream_context_create($arrContextOptions));
+            $data = json_decode($response);
+
+            // Check if any parts were actually built
+            if($data != null){
+                $output = array();
+                $i = 0;
+
+                // Store parts data into an array
+                foreach($data as $record){
+                    $output[$i]['part_ca'] = $record->{'id'};
+                    $output[$i]['part_code'] = $record->{'model'} . $record->{'piece'};
+                    $output{$i}['built_at'] = $record->{'plant'};
+                    $output{$i}['date_built'] = $record->{'stamp'};
+                    $i++;            
+                }
+
+                // Insert built parts into the database
+                $res = $this->partsmodel->insert($output);
+            }
+
+            redirect('Parts');
         }
+        else{
+            redirect('Parts');
+        }        
+    }
+
+    public function buy(){
+        // Get API key from database
+        $key = $this->apikeymodel->getKey()[0]['apikey'];
+        if($key != null){
+            $arrContextOptions=array(
+                "ssl"=>array(
+                    "verify_peer"=>false,
+                    "verify_peer_name"=>false,
+                ),
+            ); 
+
+            // Get json response from REST API
+            $response = file_get_contents('https://umbrella.jlparry.com/work/buybox?key=' . $key, false, stream_context_create($arrContextOptions));
+            $data = json_decode($response);
+
+            $output = array();
+            $i = 0;
+
+            // Store parts data into an array
+            foreach($data as $record){
+                $output[$i]['part_ca'] = $record->{'id'};
+                $output[$i]['part_code'] = $record->{'model'} . $record->{'piece'};
+                $output{$i}['built_at'] = $record->{'plant'};
+                $output{$i}['date_built'] = $record->{'stamp'};
+                $i++;            
+            }
+
+            // Insert bought parts into the database
+            $res = $this->partsmodel->insert($output);
+
+            redirect('Parts');
+        }
+        else{            
+            redirect('Parts');
+        }  
+    }
 }