+++ application/views/parts.php
@@ -2,8 +2,10 @@
    <section class="wrapper">
       <div class="col-lg-9 main-chart">      
         <h2>Parts</h2>    
-        <button>Build More Parts</button>  
-        <button>Buy Parts</button>   
+        <form>
+        <button type="submit" formaction="/Parts/Build">Build More Parts</button>  
+        <button type="submit" formaction="/Parts/Buy">Buy Parts</button>   
+        </form>
         {partsTable}
       </div>
    </section>